Add a ListItemView class
Summary: This adds a `SIMUIListItemView` class for rendering individual list items. Actually coming up with the individual menu items to render is still a little bit messy, and should probably be derived based on some global state or whatever. But this is a working prototype and I'm excited about it.

Test Plan:
Ran locally

{F175055}

Reviewers: brooke

Reviewed By: brooke

Differential Revision: https://phabricator.tbmh.org/D9