Adds ConfigurationLocalSource class
Summary:
This adds an additional class for reading and writing local configuration values.
My rough plan here is to use this for some initial configuration and setup scripts/checks
so that we can walk users through installation and configuration a little more reasonably.
(Like, check that database credentials are set and the database is reachable; set a base
URI; etc.)

Immediate next step will probably be to build that setup user flow and a utility to allow
users to programmatically interact with configuration sources, e.g., via
```
${app_root}/bin/config set mysql.host 127.0.0.1
```
or similar

Test Plan:
Just code review for now. Verified that all the JSON read/write ops work
as expected

Reviewers: brooke

Reviewed By: brooke

Differential Revision: https://phabricator.tbmh.org/D15