Correct check_database_reachable() setup check
Summary: This corrects a bug in how I was checking for database connectivity on startup. For simplicity, I just changed from an array of connections to a single connection. Down the road when we build out database clustering this'll be something we need to revisit, but it's safe to ignore for now.

Test Plan:
```
lang=python
>>> import simplex
>>> from simplex.application.config import setup_check
>>> _tmp = setup_check.SimplexDatabaseSetupCheck(simplex.SIMPLEX_ENV)
>>> _tmp.issues
[]
```

Reviewers: brooke

Reviewed By: brooke

Differential Revision: https://phabricator.tbmh.org/D22