Initial user registration workflow
Summary: This is a first stab at getting a workflow set up to configure the initial administrative user account on a fresh install. Currently, the whole "register an account" functionality is missing and it'll just fatal, but I'm planning to build that out pretty soon here 😬

Test Plan: Ran locally and saw the application crash and burn when it realized I didn't have an account. Then added the appropriate redirect and saw things get redirected to the approriate `/auth/register` route.

Reviewers: brooke

Reviewed By: brooke

Differential Revision: https://phabricator.tbmh.org/D25